backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
black==25.9.0
boto3==1.40.50
botocore==1.40.50
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
Deprecated==1.2.18
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.110.1
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.1.0
jmespath==1.0.1
jq==1.10.0
limits==5.6.0
lxml==5.3.0
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
mypy==1.18.2
mypy_extensions==1.1.0
numpy==2.3.3
oauthlib==3.3.1
openpyxl==3.1.2
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pytest==8.4.2
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
redis==7.0.0
reportlab==4.0.7
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
s3transfer==0.14.0
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.37.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.25.0
watchfiles==1.1.0
websockets==12.0
wrapt==1.17.3
